ADDED: Function to Sort Data As Per Largest State by Area Store into Json format

+++ src/main/java/com/analyser/dao/CensusDAO.java
@@ -5,7 +5,7 @@ import com.analyser.dto.StateCensusCsv;
 public class CensusDAO {
     public String state;
     public int population;
-    public double areaInSqKm;
+    public long areaInSqKm;
     public double densityPerSqKm;
     public String stateCode;
